+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5V453CTD2BREMEZ092PACZ"
+        "01KJ60BRV8CK733MJ1QJPENW9Q"
     )
     r"""The unique identifier of the data source"""