+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5V453EY56535GBNYP51MTJ"
+        "01KJ60BRVBFX7J70FR9B7C3XMZ"
     )
     r"""The unique identifier of the data source"""